Configure ssh security group rule in the demo
This change adds a second rule to the demo security group to permit
inbound ssh from the host on which the playbook is run. This removes the
need to have a pre-existing security group that works with SSH.

It also neatly demos adding multiple rules to the same security group
(the number one FAQ).

Also in htis change I was able to remove the work around for an early
ansible bug which required at least one entry in your inventory file to
use playbooks. That removes two source files and a few more lines of
crap to explain when presenting :-)

Finally I removed the code to specify a network since this isn't needed
unless you have multiple networks. Saving lines of code is critical when
presenting this demo.