Work around slow/broken floating IP assignment
Shade seems to trust that nova is in sync with neutron WRT floating IPs
and that is not the case in the system under test. This change uses
neutron to find the floating IP that was allocated. It's not pretty, but
it is much faster than even the previous approach!